The fuck am i doing

+++ src/constants.rs
@@ -8,11 +8,11 @@ pub const ASSETS_PATH: &str = "/Users/curzel/dev/tower-defense/assets";
 pub const NO_PARENT: u32 = 0;
 pub const INFINITE_LIFESPAN: f32 = -420.0;
 pub const INITIAL_CAMERA_VIEWPORT: Rectangle = Rectangle::new(0.0, 0.0, 1000.0, 750.0);
-pub const TILE_SIZE: f32 = 16.0;
-pub const TILE_MARGIN: f32 = TILE_SIZE / 16.0;
-pub const TILE_TEXTURE_SIZE: f32 = 16.0;
+pub const TILE_SIZE: i32 = 16;
+pub const TILE_MARGIN: i32 = TILE_SIZE / 16;
+pub const TILE_TEXTURE_SIZE: i32 = 16;
 pub const TILE_VARIATIONS_COUNT: u32 = 4;
-pub const COLLISION_THRESHOLD: f32 = TILE_SIZE / 3.0;
+pub const COLLISION_THRESHOLD: f32 = TILE_SIZE as f32 / 3.0;
 pub const HERO_ENTITY_ID: u32 = 69;
 pub const INFINITE_STOCK: i32 = -420;
